pyproject.toml
[project]
name = "glance"
version = "0.1.0"
description = "Multi-agent AI code review system for GitHub PRs"
requires-python = ">=3.10"
dependencies = [
    "PyGithub>=2.0",
    "python-dotenv>=1.0",
    "httpx>=0.25",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "openai>=1.0",
    "pytest>=7.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.23",
]
rag = [
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
]
local = [
    "numpy>=1.24",
    "scikit-learn>=1.3",
]

[project.scripts]
glance = "glance.orchestrator:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]
ignore = ["E501"]
